hw1-bundle/code/hw1/pyproject.toml
[project]
name = "hw1"
version = "0.1.0"
description = ""
authors = [
    {name = "pandelis-giankulidis",email = "pantelis.giankoulidis+1@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pandas (>=2.3.3,<3.0.0)",
    "pyspark (>=4.1.1,<5.0.0)",
    "pillow (>=12.1.0,<13.0.0)",
    "matplotlib (>=3.10.8,<4.0.0)"
]

[tool.poetry]
packages = [{include = "hw1", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
